// libs/services/common/src/lib/enums/claim-keys.enum.ts
export enum ClaimKeys {
  SUBJECT = 'sub',
  ISSUER = 'iss',
  AUDIENCE = 'aud',
  EXPIRATION = 'exp',
  ISSUED_AT = 'iat',
  JWT_ID = 'jti',
  ROLES = 'roles',
  TENANT_ID = 'tid',
  FULL_NAME = 'name',
  FIRST_NAME = 'given_name',
  LAST_NAME = 'family_name',
  USERNAME = 'username',
  EMAIL = 'email',
  ACCOUNT_ID = 'accid',
  CLIENTS = 'cli'
}